+++ CHANGES.md
@@ -26,6 +26,8 @@
   indented less (#3964)
 - Multiline list and dict unpacking as the sole argument to a function is now also
   indented less (#3992)
+- In f-string debug expressions preserve quote types that are visible in the final
+  string (#4005)
 - Fix a bug where long `case` blocks were not split into multiple lines. Also enable
   general trailing comma rules on `case` blocks (#4024)
 - Keep requiring two empty lines between module-level docstring and first function or